+++ app/src/main/java/com/example/myrestaurant/Model/Response/Size.java
@@ -11,7 +11,7 @@ public class Size {
 	private String description;
 
 	@SerializedName("extraPrice")
-	private float extraPrice;
+	private double extraPrice;
 
 	public int getId() {
 		return id;
@@ -29,11 +29,11 @@ public class Size {
 		this.description = description;
 	}
 
-	public float getExtraPrice() {
+	public double getExtraPrice() {
 		return extraPrice;
 	}
 
-	public void setExtraPrice(float extraPrice) {
+	public void setExtraPrice(double extraPrice) {
 		this.extraPrice = extraPrice;
 	}
